Add Wordle context and useWordle hook, refactor solution and words fetch logic, fix Keyboard and info boxes styles
