+++ components/WordleGame/Wordle.tsx
@@ -1,103 +1,12 @@
-import { Dispatch, SetStateAction, useEffect, useState } from "react";
-import formatWord from "../../helpers/formatWord";
-import { gameStateEnum, IWord, wordArr } from "../../types/types";
+import { useContext, useEffect } from "react";
+import WordleContext from "../../context/WordleContext";
+import { gameStateEnum } from "../../types/types";
 import Line from "./Line/Line";
 import styles from "./Wordle.module.css";
-import { toast } from "react-toastify";
-import checkWord from "../../helpers/checkWord";
 
-interface Props {
-  words: IWord[];
-  setWords: Dispatch<SetStateAction<IWord[]>>;
-  solution: string;
-  gameState: gameStateEnum;
-  setGameState: Dispatch<SetStateAction<gameStateEnum>>;
-}
-
-const WordleGame = ({
-  words,
-  setWords,
-  solution,
-  gameState,
-  setGameState,
-}: Props) => {
-  const [turn, setTurn] = useState<number>(0);
-  const [word, setWord] = useState<IWord>({ wordStr: "", wordArr: [] });
-  const toastId = 1;
-
-  const addWord = (word: string, idx: number) => {
-    // Set word at correct index in words array
-    setWords((prev) => {
-      return prev.map((e: IWord, i: number) =>
-        i === idx ? { wordStr: word, wordArr: formatWord(word, solution) } : e
-      );
-    });
-  };
-
-  const handleKeyUp = async (e: KeyboardEvent): Promise<void> => {
-    const key: string = e.key.toLowerCase();
-    const isSingleLetter: boolean = new RegExp(/^[a-z]$/).test(key);
-    if (!isSingleLetter && key !== "backspace" && key !== "enter") return;
-
-    if (isSingleLetter) {
-      if (word.wordStr.length < 5) {
-        const wordStr: string = word.wordStr + key;
-        const wordArr: wordArr = formatWord(wordStr, solution);
-        setWord({ wordStr, wordArr });
-      }
-    } else if (key === "backspace") {
-      if (word.wordStr.length) {
-        const wordStr: string = word.wordStr.slice(0, -1);
-        const wordArr: wordArr = formatWord(wordStr, solution);
-        setWord({ wordStr, wordArr });
-      }
-    } else if (key === "enter") {
-      if (word.wordStr.length === 5) {
-        if (word.wordStr === solution) {
-          addWord(word.wordStr, turn);
-          setWord({ wordStr: "", wordArr: [] });
-          setTurn((prev) => prev + 1);
-          setGameState(gameStateEnum.finishWin);
-          return;
-        }
-
-        const exists = (await checkWord(word.wordStr)).exists;
-        if (!exists) {
-          toast.error("Word do not exist in dictionary!", {
-            toastId: toastId,
-            position: "bottom-center",
-            autoClose: 2500,
-            hideProgressBar: false,
-            closeOnClick: true,
-            pauseOnHover: false,
-            draggable: false,
-            progress: undefined,
-            theme: "light",
-          });
-          return;
-        }
-
-        // If last turn and word is not a solution - game over
-        if (turn === 4) {
-          setGameState(gameStateEnum.finishLose);
-        }
-
-        // Check if current word is already guessed
-        const isDuplicate: boolean = words.find(
-          (_word) => _word?.wordStr === word.wordStr
-        )
-          ? true
-          : false;
-        if (isDuplicate) {
-          return;
-        }
-
-        addWord(word.wordStr, turn);
-        setWord({ wordStr: "", wordArr: [] });
-        setTurn((prev) => prev + 1);
-      }
-    }
-  };
+const WordleGame = () => {
+  const { word, words, turn, gameState, handleKeyUp } =
+    useContext(WordleContext);
 
   useEffect(() => {
     gameState === gameStateEnum.inProgress &&
