+++ helpers/checkWord.ts
@@ -5,7 +5,7 @@ interface IResponse {
 }
 
 const checkWord = async (word: string): Promise<IResponse | undefined> => {
-  const res = await axios.get(`http://localhost:5400/check?word=${word}`);
+  const res = await axios.get(`https://wordle-db.onrender.com/check?word=${word}`);
   const data = await res.data;
   return data;
 };
