+++ pages/_app.tsx
@@ -1,19 +1,22 @@
 import { AppProps } from "next/app";
-import "../styles/globals.css";
 import { QueryClientProvider, QueryClient } from "@tanstack/react-query";
+import { WordleProvider } from "../context/WordleContext";
 import { ThemeProvider } from "../context/ThemeContext";
 import Layout from "../components/Layout";
+import "../styles/globals.css";
 
 const MyApp = ({ Component, pageProps }: AppProps) => {
   const client = new QueryClient();
 
   return (
     <QueryClientProvider client={client}>
-      <ThemeProvider>
-        <Layout>
-          <Component {...pageProps} />
-        </Layout>
-      </ThemeProvider>
+      <WordleProvider>
+        <ThemeProvider>
+          <Layout>
+            <Component {...pageProps} />
+          </Layout>
+        </ThemeProvider>
+      </WordleProvider>
     </QueryClientProvider>
   );
 };
