+++ styles/Home.module.css
@@ -1,18 +1,23 @@
 .main {
   display: flex;
   align-items: center;
-
   flex-direction: column;
 }
 
+.info-container {
+  display: grid;
+  place-content: center;
+}
+
 .info {
-  margin: 0 auto;
-  padding: 1em;
-  width: 50%;
+  margin: 0 1em;
+  padding: 1.25em;
 
   display: flex;
   flex-direction: column;
   align-items: center;
+
+  gap: 0.5em;
 
   border-radius: 8px;
   background-color: var(--info-background);
@@ -21,21 +26,19 @@
 }
 
 .error-text {
+  margin: 0;
+  text-align: center;
   color: red;
-  margin: 0.5em 0;
 }
 
 .text {
-  margin: 0.5em 0;
+  margin: 0.15em 0;
   font-weight: bold;
-
-  font-size: 1.25rem;
-  opacity: 0.75;
+  text-align: center;
 }
 
 .spinner-box {
   margin: 1em 0;
-
   display: flex;
   flex-direction: column;
   justify-content: center;
